update pacakges, update service and controller

+++ src/problems.service.ts
@@ -1,4 +1,9 @@
-import { BadRequestException, Inject, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  Logger,
+} from '@nestjs/common';
 import { CreateProblemDto } from './dto/create-problem.dto';
 import { UpdateProblemDto } from './dto/update-problem.dto';
 import { InjectModel } from '@nestjs/mongoose';
@@ -10,16 +15,39 @@ import { firstValueFrom } from 'rxjs';
 
 @Injectable()
 export class ProblemsService {
+  private readonly logger = new Logger(ProblemsService.name);
+  private isConnected = false;
   constructor(
     @InjectModel(Problem.name) private ProblemModel: Model<Problem>,
-    @Inject('USERS_SERVICE') private readonly client: ClientProxy,
+    @Inject('USERS_SERVICE') private readonly usersclient: ClientProxy,
   ) {
-    this.client.connect();
+    this.connectToClient();
   }
+
+  private async connectToClient() {
+    try {
+      await this.usersclient.connect();
+      this.isConnected = true;
+      this.logger.log('Connected to the client successfully');
+    } catch (error) {
+      this.logger.error(
+        `Error connecting to client: ${error.message}`,
+        error.stack,
+      );
+    }
+  }
+
   async create(createProblemDto: CreateProblemDto): Promise<Problem> {
     try {
+      if (!this.isConnected) {
+        this.logger.warn('Client is not connected!');
+        return;
+      }
       const user = await firstValueFrom(
-        this.client.send<string>('findByUsername', createProblemDto.creator),
+        this.usersclient.send<string>(
+          'findByUsername',
+          createProblemDto.creator,
+        ),
       );
 
       if (!user) throw new BadRequestException('User not exists');
@@ -32,21 +60,37 @@ export class ProblemsService {
   }
 
   async findAll() {
+    if (!this.isConnected) {
+      this.logger.warn('Client is not connected!');
+      return;
+    }
     const result = await this.ProblemModel.find();
     return result;
   }
 
   async findOne(id: ObjectId) {
+    if (!this.isConnected) {
+      this.logger.warn('Client is not connected!');
+      return;
+    }
     const result = await this.ProblemModel.findById(id);
     return result;
   }
 
   async update(id: ObjectId, updateProblemDto: UpdateProblemDto) {
+    if (!this.isConnected) {
+      this.logger.warn('Client is not connected!');
+      return;
+    }
     const result = await this.ProblemModel.updateOne(id, updateProblemDto);
     return result;
   }
 
   async remove(id: ObjectId): Promise<any> {
+    if (!this.isConnected) {
+      this.logger.warn('Client is not connected!');
+      return;
+    }
     const result = await this.ProblemModel.deleteOne(id);
     return result;
   }
